requirements.txt
# Telegram Bot
python-telegram-bot==21.6

# Claude API
anthropic==0.40.0

# 環境變數
python-dotenv==1.0.1

# 資料處理
pandas==2.2.0
numpy==1.26.4

# 技術指標
ta==0.11.0

# 機器學習
lightgbm==4.3.0
scikit-learn==1.4.0
joblib==1.3.2

# 爬蟲
requests==2.31.0
beautifulsoup4==4.12.3

# 排程
schedule==1.2.1

# 視覺化（選股報告圖表）
matplotlib==3.8.2
